Fixes #63
Integrated with new eureka functionality of providing a HealthcheckHandler that provides detailed eureka status as opposed to a binary value.

Since, we need to be backward compatible with existing Healthcheck implementation, this change deprecates the older way and adds a new way of healthcheck. The karyon healthcheck abstraction does not change.

The new functionality can be enabled by using the property "com.netflix.karyon.eureka.use.healthcheck.handler=true"

Also updated the example to demonstrate this behavior.